requirements.txt
requests
pydotenv